Update MetricDetail.svelte
Added Metric Info to the Metric page.

The keys currently are the same as media_state_play page.

Formatting/styling has not yet been added as mentioned in the Issue.

More info (help) has been linked for lifetime key. It can be further added for others in the similar manner.

This is for Issue : Metric pages show almost no info #66