Fix angle brackets (hopefully for the last time)
PR #497 fixed some cases, but not all. Angle brackets in
code blocks have their own escaping behaviour, which doesn't
work well with the adhoc work we're doing.

Strictly a string like this:

<provider>

is considered HTML in markdown. However this *probably* isn't
how we want to interpret it in the Glean Dictionary. We'll
address this by modifying the renderer when it encounters an
HTML tag, rather than doing a naive regex replace against the
whole markdown string.