Don't use a reactive declarations to handle updating the URL (#780)
It's unreliable during page navigation and often causes us to persist
unwanted state and/or weird bugs. The basic problem is that navigation
(and the subsequent updating of page state) can trigger a reactive
statement inside a component we're navigating *away* from.

Instead, let's explicitly make a call to update the state when we need
to do this, which will also update the URL. It's a little less elegant
but at least it seems to work.