Attempt to modularize build-glean-metadata script
In particular, seperate out etl logic dealing with bigquery, looker,
and glam etl into their own files. This will hopefully make it easier
to understand, extend and write tests on in the future.

This change deliberately tries to keep the output as similar as possible
to what it was before, to make it easier to audit the output.