Better breadcrumbs and titles (#1033)
Use the *actual* data for the applications, metrics, etc. in the
titles and breadcrumbs, not the normalized values derived from the URL.

This also fixes a few bugs around when/where to apply the android/ios
tags to the breadcrumbs by using the app metadata instead of inferring
it from the title.