Handle missing `looker` property
`pingData.looker` might be null, when there's no Looker explore we know
about.
This in turn made this whole expression explode, because `e[3].looker is
undefined`.
This results in a blank page, see e.g. any metric on the VPN app.

Instead we just pass on `undefined`. `getSTMOQueryURL` only needs that
parameter for events and that code path already handles this value being
falsy.