Glean ETL: Pick the latest known ping data
There's instances where the ping was formerly defined in a dependency
and was then moved to the app itself.
The app definition is what should matter now.

What if it happened the other way around?
We have history for that!
Pings should never be defined twice, so we can use history of both to
decide.
The last touched one is the correct info.

We do _not_ merge histories, so we do lose _some_ information, like when
it was first added.